+++ client/src/hooks/Conversations/index.ts
@@ -2,4 +2,3 @@ export { default as usePresets } from './usePresets';
 export { default as useGetSender } from './useGetSender';
 export { default as useDebouncedInput } from './useDebouncedInput';
 export { default as useParameterEffects } from './useParameterEffects';
-export { default as useImportFileHandling } from './useImportFileHandling';
